pyproject.toml
[project]
name = "docmd"
version = "0.1.0"
description = "Convert markdown files to docx or pdf using docxtpl templates"
readme = "README.md"
authors = [
    { name = "renatormc", email = "renatomartinsrmc@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "docxtpl[subdoc]>=0.20.2",
]

[project.scripts]
docmd = "docmd.__main__:main"

[dependency-groups]
dev = [
    "mypy>=1.20.2",
    "pytest>=8.0",
]

[build-system]
requires = ["uv_build>=0.10.8,<0.11.0"]
build-backend = "uv_build"
